% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\title{An App to Compare Two Data Frames}
\usage{
compare_df(df_left, df_right, df_names = NULL)
}
\arguments{
\item{df_left}{\code{<tbl|df>} Left tibble or data frame}

\item{df_right}{\code{<tbl|df>} Right tibble or data frame}
}
\description{
An App to Compare Two Data Frames
}
