% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{diff_app}
\alias{diff_app}
\title{An App to Compare Two Data Frames}
\usage{
diff_app(x, y, df_names = NULL)
}
\arguments{
\item{x}{\code{<tbl|df>} Left tibble or data frame}

\item{y}{\code{<tbl|df>} Right tibble or data frame}
}
\description{
An App to Compare Two Data Frames
}
