% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{diff_cols_pmatch}
\alias{diff_cols_pmatch}
\title{Use Partial Matching to Find Common Column Names}
\usage{
diff_cols_pmatch(x, ...)
}
\arguments{
\item{x}{\code{<diff_tbl|tbl>} \code{diff_tbl} output from \code{\link[=diff_compare]{diff_compare()}} or the
reference data frame or tibble.}

\item{...}{Arguments passed on to \code{diff_pair}
\describe{
  \item{x}{\code{<tbl|matrix>} A data set (\code{tibble}, \code{data.frame}, \code{matrix}) or a
a (named) list containing 2 data sets.}
}}
}
\description{
Use Partial Matching to Find Common Column Names
}
\seealso{
Other Column match resolvers: \code{\link{diff_cols_common}},
  \code{\link{diff_cols_type}},
  \code{\link{diff_cols_unique}}
}
\concept{Column match resolvers}
