% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{diff_cols_unique}
\alias{diff_cols_unique}
\title{Unique columns}
\usage{
diff_cols_unique(x, ...)
}
\arguments{
\item{x}{\code{<diff_tbl|tbl>} \code{diff_tbl} output from \code{\link[=diff_compare]{diff_compare()}} or the
reference data frame or tibble.}

\item{...}{Arguments passed on to \code{diff_compare}
\describe{
  \item{x}{\code{<tbl|df>} Reference data frame}
  \item{y}{\code{<tbl|df>} Comparison data frame}
  \item{exclude}{\code{<chr:NULL>} ID column(s) that are excluded from the
differencing}
  \item{keys}{\code{<chr:ignore>} Column names of grouping variables for
alignment, takes default from \code{ignore}.}
  \item{align}{\code{<lgl:FALSE>} Should alignmment be performed based on the
grouping variables or the \code{group_vars} parameter?}
  \item{df_names}{\code{<chr:NULL>} Alternative names for the provided dataframes as
a vector of length two, to be used during printing and reporting.}
  \item{tolerance}{\code{<dbl>} The tolerance to be used for comparison of numerical
values, defaults to \code{.Machine$double.eps}.}
  \item{plain}{\code{<lgl:FALSE>} If \code{TRUE}, returns a nested tibble of differences
without converting these differences to a \code{diff_tbl} object.}
}}
}
\description{
Unique columns
}
\seealso{
Other Column match resolvers: \code{\link{diff_cols_common}},
  \code{\link{diff_cols_match}},
  \code{\link{diff_cols_pmatch}},
  \code{\link{diff_cols_type}}
}
\concept{Column match resolvers}
