% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_pull.R
\name{diff_pull}
\alias{diff_pull}
\title{Pull Out a Variable}
\usage{
diff_pull(x, var = NULL, regexp = NULL)
}
\arguments{
\item{x}{\code{<diff_tbl|tbl>} \code{diff_tbl} output from \code{\link[=diff_compare]{diff_compare()}} or the
reference data frame or tibble.}

\item{var}{\code{<chr>} Vector of variable names}

\item{regexp}{\code{<chr>} Regular expression used as a filter to find variable
names. If provided, this filter is applied first, and only columns matching
the \code{regexp} and appearing in the \code{var} list are included.}
}
\description{
Pull Out a Variable
}
