% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{diff_report}
\alias{diff_report}
\title{Differences Report}
\usage{
diff_report(reference, comparison, df_names = NULL, outfile = NULL,
  keep_original = FALSE, use_plotly = is.null(outfile),
  use_DT = is.null(outfile), quiet = TRUE, ...)
}
\arguments{
\item{reference}{\code{<tbl>} Reference data frame or tibble}

\item{comparison}{\code{<tbl>} Comparison data frame or tibble}

\item{df_names}{\code{<chr>} Vector of length two containg the names of the data
frames that should be displayed in the report}

\item{outfile}{\code{<chr>} Filename (with optional path) to which report should be
saved. By default writes to temp file and opens in RStudio Viewer pane.}

\item{keep_original}{\code{<lgl>} Should the original data frames be included
in the report?}

\item{use_plotly}{\code{<lgl>} Should the differences visualization be interactive
via \link{plotly}? By default this is \code{FALSE} when saving the report to a
standalone file.}

\item{use_DT}{\code{<lgl>} Should the tables reporting differences be presented as
an interactive datatable with the \link{DT} package? By default this is \code{FALSE}
when saving the report to a standalone file.}

\item{quiet}{\code{TRUE} to suppress printing of the
  pandoc command line.}

\item{...}{Arguments passed on to \code{diff_compare}
\describe{
  \item{ignore}{\code{<chr:NULL>} ID column(s) that are ignored from the differencing}
  \item{group_vars}{\code{<chr:ignore>} Column names of grouping variables for alignment,
takes default from \code{ignore}.}
  \item{align}{\code{<lgl:FALSE>} Should alignmment be performed based on the
grouping variables or the \code{group_vars} parameter?}
  \item{tolerance}{\code{<dbl>} The tolerance to be used for comparison of
numerical values, defaults to \code{.Machine$double.eps}.}
}}
}
\value{
Invisibly returns path to rendered differences report
}
\description{
Creates an HTML report for exploring the differences between two data frames
or tibbles. If \code{outfile} is specified, the report is saved as a standalone
HTML file that can be shared with others.
}
