% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/different-package.R
\docType{package}
\name{different-package}
\alias{different}
\alias{different-package}
\title{different: Compare Differences in Data Frames}
\description{
Tools for comparing and resolving differences between data
        frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GerkeLab/different}
  \item Report bugs at \url{https://github.com/GerkeLab/different/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}

}
\keyword{internal}
