% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_add_ids}
\alias{dummy_add_ids}
\title{Add IDs to Dummy Data Core}
\usage{
dummy_add_ids(df = dummy_data_core(), n_id_col = 2, id_names = NULL)
}
\arguments{
\item{df}{\code{<tbl>} Input data frame to which ID collumns shold be addeed}

\item{n_id_col}{\code{<int>} Number of ID columns to add}

\item{id_names}{\code{<chr>} Vector of names for ID columns}
}
\description{
Add IDs to Dummy Data Core
}
