% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_col}
\alias{dummy_col}
\alias{dummy_chr}
\alias{dummy_fct}
\alias{dummy_int}
\alias{dummy_dbl}
\title{Dummy Column}
\usage{
dummy_col(n, type = c("int", "dbl", "chr", "fct"), ...)

dummy_chr(n, chr_len = 6)

dummy_fct(n, fct_n_levels = round(n * 0.333, 0))

dummy_int(n, int_min = -50L, int_max = 50L)

dummy_dbl(n, dbl_rfun = function(n, ...) {     runif(n, 0, 1) *
  10^sample(1:5, 1) }, ...)
}
\arguments{
\item{n}{Number of elements}

\item{type}{One of \code{"int"}, \code{"dbl"}, \code{"chr"} or \code{"fct"}}

\item{...}{Arguments passed to associated helper functions. All helper
\code{dummy_col()} functions ignore extra parameters, except for \code{dummy_dbl()}
where additional parameters are passed to \code{dbl_rfun}.}

\item{chr_len}{\code{<int>} Length of strings}

\item{fct_n_levels}{\code{<int>} Number of factor levels}

\item{int_min}{\code{<int>} Minimum integer value}

\item{int_max}{\code{<int>} Maximum integer value}

\item{dbl_rfun}{\code{<fun>} A function that takes an argument \code{n} and generates
a vector of \code{n} random doubles.}
}
\description{
Dummy Column
}
\examples{
dummy_chr(10, chr_len = 4)
dummy_fct(10, fct_n_levels = 5)
dummy_int(10)
dummy_int(10, int_min = 1, int_max = 3)
dummy_dbl(10)
dummy_dbl(10, runif, min = 1, max = 10)
}
