% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_corrupt}
\alias{dummy_corrupt}
\title{Corrupt Dummy Data}
\usage{
dummy_corrupt(df = dummy_data(), ..., n_corrupted = nrow(df)/5,
  rx_exclude = "id|ID", shuffle = TRUE, corrupt_type_rate = 0.25,
  dropout = 0.1, dropout_col = dropout, dropout_row = dropout,
  paired = TRUE)
}
\arguments{
\item{df}{\code{<tbl>} Data frame to be corrupted}

\item{...}{Columns to be included for corruption, unquoted following
\code{\link[dplyr:select]{dplyr::select()}} semantics.}

\item{n_corrupted}{\code{<int>} Number of rows to be corrupted within each
selected column.}

\item{rx_exclude}{\code{<regexp>} Regular expression indicating which columns
should be protected from corruption. Set to \code{NULL} to disable.}

\item{shuffle}{\code{<lgl>} If \code{TRUE} (default), rows are suffled}

\item{corrupt_type_rate}{\code{<dbl>} Rate between 0 and 1 at which column type
should be corrupted. A rate of 0.25 means that each column will have a 25%
chance of being corrupted.}

\item{dropout}{\code{<dbl>} Percentage of rows/columns to drop (default value of
\code{dropout_col} or \code{dropout_row} unless those values are otherwise
specified.)}

\item{dropout_col}{\code{<dbl>} Percentage of columns to drop}

\item{dropout_row}{\code{<dbl>} Percentage of rows to drop}

\item{paired}{\code{<lgl>} If \code{TRUE} (default), returns a \code{\link[=diff_pair]{diff_pair()}} with the
original and corrupted data sets.}
}
\description{
Corrupt Dummy Data
}
