% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_data}
\alias{dummy_data}
\title{Create Dummy Data}
\usage{
dummy_data(n_row = 100, n_col = 10, types = sample(c("int", "dbl",
  "fct", "chr"), n_col, replace = TRUE), n_id_col = 2, id_names = NULL,
  ...)
}
\arguments{
\item{n_row}{\code{<int>} Number of rows}

\item{n_col}{\code{<int>} Number of columns}

\item{types}{\code{<chr>} Data types of the columns, recycled or truncated to
match \code{n_col}, or by default chosen at random. If named, column names will
match the names of this vector.}

\item{n_id_col}{\code{<int>} Number of ID columns to add}

\item{id_names}{\code{<chr>} Vector of names for ID columns}
}
\description{
Create Dummy Data
}
