% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_data_core}
\alias{dummy_data_core}
\title{Generate Core Dummy Data}
\usage{
dummy_data_core(n_row = 100, n_col = 10, types = sample(c("int",
  "dbl", "fct", "chr"), n_col, replace = TRUE), ...)
}
\arguments{
\item{n_row}{\code{<int>} Number of rows}

\item{n_col}{\code{<int>} Number of columns}

\item{types}{\code{<chr>} Data types of the columns, recycled or truncated to
match \code{n_col}, or by default chosen at random. If named, column names will
match the names of this vector.}

\item{...}{Arguments passed on to \code{dummy_col}
\describe{
  \item{n}{Number of elements}
  \item{type}{One of \code{"int"}, \code{"dbl"}, \code{"chr"} or \code{"fct"}}
}}
}
\description{
Generate Core Dummy Data
}
