% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{fuzzy_match_str}
\alias{fuzzy_match_str}
\title{Fuzzy string matching}
\usage{
fuzzy_match_str(x, y, method = c("word", "bigram", "stringdist"),
  stringdist_args = list(method = "cosine"))
}
\description{
Goal: Take (column names as) two vectors of string, return tibble with
matching candidates.
}
