% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{diff_report}
\alias{diff_report}
\title{Differences Report}
\usage{
diff_report(x, y = NULL, ..., df_names = NULL, output_file = NULL,
  keep_original = FALSE, use_plotly = is.null(output_file),
  use_DT = is.null(output_file), notes = NULL, quiet = TRUE)
}
\arguments{
\item{x}{\code{<diff_tbl|tbl>} \code{diff_tbl} output from \code{\link[=diff_compare]{diff_compare()}} or the
reference data frame or tibble.}

\item{...}{Arguments passed on to \code{diff_compare}
\describe{
  \item{exclude}{\code{<chr:NULL>} ID column(s) that are excluded from the
differencing}
  \item{keys}{\code{<chr:ignore>} Column names of grouping variables for
alignment, takes default from \code{ignore}.}
  \item{align}{\code{<lgl:FALSE>} Should alignmment be performed based on the
grouping variables or the \code{group_vars} parameter?}
  \item{tolerance}{\code{<dbl>} The tolerance to be used for comparison of numerical
values, defaults to \code{.Machine$double.eps}.}
}}

\item{df_names}{\code{<chr>} Vector of length two containg the names of the data
frames that should be displayed in the report}

\item{output_file}{\code{<chr>} Filename (with optional path) to which report should
be saved. By default writes to temp file and opens in RStudio Viewer pane.}

\item{keep_original}{\code{<lgl>} Should the original data frames be included in
the report?}

\item{use_plotly}{\code{<lgl>} Should the differences visualization be interactive
via \link{plotly}? By default this is \code{FALSE} when saving the report to a
standalone file.}

\item{use_DT}{\code{<lgl>} Should the tables reporting differences be presented as
an interactive datatable with the \link{DT} package? By default this is \code{FALSE}
when saving the report to a standalone file.}

\item{notes}{\code{<chr>} A character vector containing notes in markdown format
to be embedded in the report.}

\item{quiet}{\code{TRUE} to suppress printing of the
  pandoc command line.}

\item{.y}{\code{<tbl>} Comparison data frame or tibble}

\item{.x}{\code{<tbl>} Original reference data frame or tibble, provided only if
\code{x} is a \code{diff_tbl}.}
}
\value{
Invisibly returns path to rendered differences report
}
\description{
Creates an HTML report for exploring the differences between two data frames
or tibbles. If \code{outfile} is specified, the report is saved as a standalone
HTML file that can be shared with others.
}
